require: improved module resolution (#2)
Trying as much as possible to avoid I/O by resolving module URIs via a
resolver map that is provided.
Therefore we fall back to Node.js resolve mechanism only when all else
fails. Additionally we try to use the Node.js module cache as well as
snapshot caches to avoid loading the module from disk as much as
possible.

The update of those caches has been improved as well.
As an aside a minor sourcemap fix was applied as well.

* resolver: cache:direct resolving cached modules

* resolver: cache:direct resolving module definitions
  - refactored init module code to ease reuse

* resolver: further optimized module resolution

* resolver: updating Module._cache with resolved modules

* resolution: updating caches when module resolved
  - keeping both Node.js and customRequire caches up to date where
    possible

* resolution: fix circular resolution handling

* sourcemap: no longer crashing when no codeFrames from `wrapCallSite`

* resolution: fall back to direct fullPath for tricky modules

* resolution: try cache again after Node.js resolve

* resolution: fix circular definition loading